feat: Detach approve/deny/get endpoints for pending users from instance
