+++ fluxer_admin/tests/htmx_acceptance.rs
@@ -767,7 +767,7 @@ async fn mock_api(method: Method, uri: Uri) -> Response {
         }
         (Method::POST, "/admin/jobs/get") => json_response(json!({ "job": searched_job() })),
         (Method::POST, "/admin/instance-config/get") => json_response(instance_config()),
-        (Method::POST, "/admin/instance-config/pending-registrations/get") => {
+        (Method::POST, "/admin/pending-registrations/get") => {
             json_response(pending_user_registrations())
         }
         (Method::POST, "/admin/instance-config/registration-urls/create") => json_response(json!({
@@ -778,12 +778,8 @@ async fn mock_api(method: Method, uri: Uri) -> Response {
         (Method::POST, "/admin/instance-config/registration-urls/revoke") => {
             json_response(instance_config_without_registration_urls())
         }
-        (Method::POST, "/admin/instance-config/pending-registrations/approve") => {
-            json_response(instance_config_without_pending_registrations())
-        }
-        (Method::POST, "/admin/instance-config/pending-registrations/reject") => {
-            json_response(instance_config_without_pending_registrations())
-        }
+        (Method::POST, "/admin/pending-registrations/approve") => json_response(json!([])),
+        (Method::POST, "/admin/pending-registrations/reject") => json_response(json!([])),
         (Method::POST, "/admin/limit-config/get") => json_response(limit_config()),
         _ => (StatusCode::NOT_FOUND, Json(json!({ "error": "not found" }))).into_response(),
     }
@@ -1077,12 +1073,6 @@ fn pending_registration_fixture() -> Value {
         "registration_url_id": "11111111-1111-4111-8111-111111111111",
         "client_ip": "203.0.113.24"
     })
-}
-
-fn instance_config_without_pending_registrations() -> Value {
-    let mut config = instance_config();
-    config["registration"]["pending_registrations"] = json!([]);
-    config
 }
 
 fn instance_config_without_registration_urls() -> Value {
