+++ fluxer_api/src/api/admin/tests/InstanceConfigPendingRegistrationAudit.test.ts
@@ -65,7 +65,7 @@ describe('pending registration audit logs', () => {
 			.execute();
 
 		await createBuilder(harness, admin.token)
-			.post(`/admin/instance-config/pending-registrations/${decision}`)
+			.post(`/admin/pending-registrations/${decision}`)
 			.header('X-Audit-Log-Reason', 'Registration review')
 			.body({user_id: pending.user_id})
 			.execute();
